rename base to CgEth2Base
