+++ src/renderer/services/eth2/client/eth2ApiClientV2/CgEth2Base.ts
