+++ src/renderer/services/eth2/client/eth2ApiClientV2/CgEth2NodeApi.ts
@@ -1,4 +1,4 @@
-import {CgEth2Base} from "./base";
+import {CgEth2Base} from "./CgEth2Base";
 import {
     Api,
     NetworkIdentity,
